Add repository existence to path resources in state
Because it is also one of the path resource.
e.g. fzy with no command.link section

- name: jhawthorn/fzy
  description: A better fuzzy finder
  owner: jhawthorn
  repo: fzy
  command:
    build:
      steps:
      - make
      - sudo make install

this should have command.link section to install fzy
and it's also used as recording in state
but if user is writing this kinda yaml and do not have command.link
this command will be installed again and again
because afx cannot detect whether command installation is successfully
done or not.
But if repository path existence is added to state path,
this will be regarded as installation success.